# src/main/resources/static/openapi.yaml
openapi: 3.0.1
info:
  title: Air-quality-hackathon-project
  description: >
    air-quality-hackathon-project
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
tags:
  - name: Air Quality
    description: Everything about Air Quality
  - name: User
    description: Everything about users
paths:

  '/api/v1/users':
    post:
      tags: [ User ]
      summary: Create user
      description: This can only be done by the logged in user.
      operationId: createUser
      requestBody:
        description: Created user object
        content:
          'application/json':
            schema:
              $ref: '#/components/schemas/CreateUserRequest'
        required: true
      responses:
        default:
          description: successful operation
          content: { }


  '/api/v1/users/me':
    get:
      tags: [ User ]
      summary: Get the current user
      operationId: readUser
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'

    put:
      tags: [ User ]
      summary: Updated user password
      description: This can only be done by the logged in user.
      operationId: updatePassword
      requestBody:
        description: Updated user object
        content:
          'application/json':
            schema:
              $ref: '#/components/schemas/ChangePasswordRequest'
        required: true
      responses:
        200:
          description: ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'

    delete:
      tags: [ User ]
      summary: Delete user
      description: This can only be done by the logged in user.
      operationId: deleteUser
      responses:
        200:
          description: ok
          content: { }
  '/api/v1/users/me/settings':
    get:
      tags: [ User ]
      summary: Get the current settings
      operationId: readSettings
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserSettingsDTO'

    put:
      tags: [ User ]
      summary: Updated user settings
      description: This can only be done by the logged in user.
      operationId: updateSettings
      requestBody:
        description: Updated user settings
        content:
          'application/json':
            schema:
              $ref: '#/components/schemas/UserSettingsDTO'
        required: true
      responses:
        200:
          description: ok
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserSettingsDTO'

  '/api/v1/air-quality':
    get:
      tags: [ Air Quality ]
      summary: Get the current Air Quality
      operationId: readAirQualityData
      responses:
        200:
          description: successful operation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AirQualityData'
components:
  schemas:
    AirQualityData:
      type: object

    UserSettingsDTO:
      type: object
      required: [emailNotificationPeriod]
      properties:
        emailNotificationPeriod:
          type: string
          enum: [ NEVER, THIRTY_MIN, ONE_HOUR, TWO_HOURS, THREE_HOURS, ONE_TIME_PER_DAY]
          default: NEVER
        stationNames:
          type: array
          items:
            type: string
        quietHoursStart:
          type: string
          pattern: '^(([0]?[1-9]|1[0-2])(:)([0-5][0-9]))$'
        quietHoursEnd:
          type: string
          pattern: '^(([0]?[1-9]|1[0-2])(:)([0-5][0-9]))$'

    ChangePasswordRequest:
      type: object
      properties:
        currentPassword:
          format: password
          type: string
        newPassword:
          format: password
          type: string
        newPasswordAgain:
          format: password
          type: string

    CreateUserRequest:
      type: object
      properties:
        username:
          type: string
        email:
          format: email
          type: string
        password:
          format: password
          type: string
        passwordAgain:
          format: password
          type: string

    UserResponse:
      type: object
      properties:
        username:
          type: string